Moved the resolvers to a more logical directory, sinds resolvers can be anything and this is specific to the audit.php
Added the sanctum guard to the guards in the audit config so that shit is going correctly into the database.
